Added network communications.
The first time the game is started, a server is created. The second time
the server is started, the client will automatically connect to the
server. So it will only work on a single machine for now.

Messages are sent between the 2 games but only printed (game state not
synchronized)